xhci: add GRUB driver glue
This adds grub-core/bus/usb/xhci_driver.c, the code that connects the
GRUB USB stack to the Coreboot xHCI driver.

I'm adding a 'portno' field to the GRUB USB device structure. It's not
ideal (I want to make as little changes as possible to GRUB code), but I
need a key to map between Coreboot and GRUB. The portno field is used to
remember which device coreboot attached to in detect_dev() and reuse
that information in control_transfer().

This fixes the problem that devices weren't properly registered by GRUB
(only coreboot xhci driver) if the device was already inserted when the
driver loads.

The driver is run through astyle for consitent coding style using these
options:

  astyle --style=gnu
         --pad-paren-out
         --pad-oper
         --align-pointer=name
         --indent=spaces=2