xhci: only get descriptor one time (not 3)
There is a logic error in usb.c::get_descriptor() that cause the
dev->controller->control function (e.g. xhci_control) to be invoked 3
times (GET_DESCRIPTOR_TRIES) even if it succeeds the first time. Fix
it.

Ensure that on success we (still) return the number of bytes we
received.