xhci: always store Max Packet Size for the xHC
Always store the Max Packet Size for the xHC (in the Input Context ->
Endpoint Context). Previously it was only stored if != 8.

This change fixes enumeration of USB 1.1 devices with MPS = 8. (Tested:
Akasa Smart Card Reader.)

An alternative fix is to modify usb.c::speed_to_default_mps() like this:

         case FULL_SPEED:
 +               /* 8, 16, 32, 64 are allowed, select the most compatible */
 +               return 8;
         case HIGH_SPEED:

But I think this is more correct.