grub/usb/xhci_driver: add 'portno' field to device structure
Not ideal (I want to make as little changes as possible to GRUB code),
but I had to make a key to map between Coreboot and GRUB. The portno
field is used to remember which device coreboot attached to in
detect_dev() and reuse that information in control_transfer().

This fixes the problem that devices weren't properly registered by GRUB
(only coreboot xhci driver) if the device was already inserted when the
driver loads.